Fix logging crashes in PrivateProfile files when bad apps call with NULLs
Some apps call the PrivateProfile functions with NULL for the filename.  Sometimes they seem to do this to get the default value and maybe the developers override with an ini file not in the production product? The native code returns something reasonable without an exception, but logging code in the FRF crashes the app. This set of fixes protects against bad logging.