Add new RegLegacyFixups fixup type "FakeDelete" (#150)
* Added a new config.json option for PsfLauncher

Added a new config.json option for PsfLauncher that is used for DEBUG builds only, the option signals PsfLauncher to WaitForDebugger as soon as it can read the request in the config.json file.  This will help with debugging issues with the launcher.

* Fix log crashes in PrivateProfile files when bad apps call with NULLs

Some apps call the PrivateProfile functions with NULL for the filename.  Sometimes they seem to do this to get the default value and maybe the developers override with an ini file not in the production product? The native code returns something reasonable without an exception, but logging code in the FRF crashes the app. This set of fixes protects against bad logging.

* Added FakeDelete type to RegLegacyFixups

* refresh

* Delete ArchitectureTest.appx.makeappx.txt

* Delete ArchitectureTest.appx.signtool.txt

* Delete CompositionTest.appx.makeappx.txt

* Delete CompositionTest.appx.signtool.txt

* Delete DynamicLibraryTest.appx.makeappx.txt

* Delete DynamicLibraryTest.appx.signtool.txt

* Delete FileSystemTest.appx.makeappx.txt

* Delete FileSystemTest.appx.signtool.txt

* Delete LongPathsTest.appx.makeappx.txt

* Delete LongPathsTest.appx.signtool.txt

* Delete PowershellScriptTest.appx.makeappx.txt

* Delete PowershellScriptTest.appx.signtool.txt

* Delete RegLegacyTest.appx.makeappx.txt

* Delete RegLegacyTest.appx.signtool.txt

* Delete WorkingDirectoryTest.appx.makeappx.txt

* Delete WorkingDirectoryTest.appx.signtool.txt

* remove commented code