Framework Package Support (#243)
# Why is this change being made?

We found a gap in use-case where apps looks up for installed version of a dependency by themselves. This is common in apps which require java, where apps look in registry or environment variable to determine the specific instance of java to use. 
This change allows PSF to intercept such lookups, map it to a dependency mentioned in AppxManifest and return back valid information about the dependency.


# What changed?
- Added new type of RegLegacyFixup and updated EnvVarFixup which allows users to specify what information app expects when it tries to read the registry to retrieve some information.
- Added automated testcases to verify the change.
- Updated README with relevant information.
- Add support for tracing the updated fixups.
- Fix few existing bugs around logging.